BACKLOG-17575: check page type in ContentTypeSelector
