+++ src/javascript/SelectorTypes/Picker/configs/Picker2.configs.js
@@ -11,7 +11,7 @@ import {registerMediaPickers} from './mediaPicker/mediaPicker';
 import {registerEditorialLinkPicker} from './editorialLinkPicker';
 import {registerPagePicker} from '~/SelectorTypes/Picker/configs/pagePicker';
 import {PickerSearchQueryHandler} from '~/SelectorTypes/Picker/configs/queryHandlers';
-import {registerEditorialPicker} from '~/SelectorTypes/Picker/configs/editorialPicker';
+import {registerEditorialPicker} from '~/SelectorTypes/Picker/configs/editorialPicker/editorialPicker';
 
 export const registerPickerConfig = registry => {
     registry.add(Constants.pickerConfig, 'default', mergeDeep({}, ContentPickerConfig, {
