+++ src/javascript/SelectorTypes/Picker/configs/editorialLinkPicker/EditorialLinkContentTypeSelector.jsx
@@ -5,13 +5,17 @@ import classes from './EditorialLinkContentTypeSelector.scss';
 import {useTranslation} from 'react-i18next';
 import {Constants} from '~/SelectorTypes/Picker/Picker2.constants';
 import {cePickerSetTableViewType} from '~/SelectorTypes/Picker/Picker2.redux';
+import {registry} from '@jahia/ui-extender';
 
 const localStorage = window.localStorage;
 
 const EditorialLinkContentTypeSelector = () => {
     const {t} = useTranslation('content-editor');
     const tableView = useSelector(state => state.contenteditor.picker.tableView);
+    const pickerKey = useSelector(state => state.contenteditor.picker.pickerKey);
     const dispatch = useDispatch();
+
+    const selectableTypesTable = registry.get('pickerConfiguration', pickerKey)?.selectableTypesTable || [];
 
     const {CONTENT, PAGES} = Constants.tableView.type;
     const LOCAL_STORAGE_VIEW_TYPE_KEY = 'jcontent_view_type';
@@ -21,7 +25,9 @@ const EditorialLinkContentTypeSelector = () => {
         localStorage.setItem(LOCAL_STORAGE_VIEW_TYPE_KEY, viewType);
     };
 
-    return (
+    const canSelectPages = selectableTypesTable.includes('jnt:page');
+
+    return canSelectPages && (
         <Tab className={classes.tabs}>
             <TabItem isSelected={PAGES === tableView.viewType}
                      data-cm-view-type={PAGES}
