+++ src/javascript/SelectorTypes/Picker/configs/editorialPicker/index.js
@@ -1,0 +1,4 @@
+export * from './editorialPicker';
+export {
+    EditorialContentTypeSelector
+} from '~/SelectorTypes/Picker/configs/editorialPicker/EditorialContentTypeSelector';
